redo price scraping + other additions
i replace scraping the price by getting it via "googlefinance" module. i
also add "current_percentage" and "ideal_percentage" columns to
"portfolio", reflecting respectively the percentage of the portfolio
currently accounted for by a security and the percentage that would be
accounted for in the ideal allocation.

also made some formatting changes.